♻️ Refactor View
