+++ client/app/(user)/todo/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ActivityIndicator, FlatList } from 'react-native';
 import { Link } from 'expo-router';
 import { useTodos, useUpdateTodo } from '@/api/todo.api';
-import { View } from '@/components/ui/View';
+import { View } from '@/components/ui/View/View';
 import { Text } from '@/components/ui/Text/Text';
 import { Button } from '@/components/ui/Button/Button';
 import { Card } from '@/components/ui/Card/Card';
