+++ client/app/(user)/ui-demo.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 import { ScrollView } from 'react-native';
 import { useTheme } from '@/providers/themeProvider';
-import { View } from '@/components/ui/View';
+import { View } from '@/components/ui/View/View';
 import { Text } from '@/components/ui/Text/Text';
 import { Button } from '@/components/ui/Button/Button';
 import { TextInput } from '@/components/ui/Input/Input';
