+++ client/components/ui/Header/Header.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Platform } from 'react-native';
 import { useRouter } from 'expo-router';
 import { useTheme } from '@/providers/themeProvider';
-import { View } from '../View';
+import { View } from '../View/View';
 import { Text } from '../Text/Text';
 import { Inline } from '../Layout/Layout';
 import { TouchableOpacity } from 'react-native';
