+++ client/components/ui/Input/__tests__/Input.test.tsx
@@ -10,17 +10,11 @@ import {
 } from '../Input';
 import {
   renderWithTheme,
-  renderWithThemeVariant,
   MockIcon,
-  TEST_CONSTANTS,
-  TEST_EVENTS,
   TEST_DATA, // Backward compatibility
   INPUT_SCENARIOS,
   ACCESSIBILITY_HELPERS,
   INPUT_STATE_HELPERS,
-  THEME_HELPERS,
-  TEST_PATTERNS,
-  createTestThemeContext
 } from './Input.test.utils';
 
 describe('<TextInput />', () => {
