+++ client/components/ui/index.ts
@@ -1,6 +1,6 @@
 // Themed styled components - export all themed components
 export * from './Text/Text';
-export * from './View';
+export * from './View/View';
 export * from './Button/Button';
 export * from './Input/Input';
 export * from './Card/Card';
